basic functionality of feateres presenter was developed
