+++ src/presenter/popup-presenter.js
@@ -6,89 +6,73 @@ import {
   RenderPosition
 } from "../utils/render-utils";
 
+import PopupFeaturesPresenter from "./popup-features-presenter";
 import CommentsPresenter from "./comments-presenter";
 import NewCommentPresenter from "./new-comment-presenter";
 
 export default class PopupPresenter {
-  constructor(handleChangeData, popupState) {
+  constructor(popupState, handleChangeData) {
     this._popupContainerElement = document.body;
+
     this._handleChangeData = handleChangeData;
+
     this._popupComponent = null;
     this._prevPopupComponent = null;
     this._newCommentPresenter = null;
-    this._isNewCommentRendered = false;
     this._popupState = popupState;
 
     this._handlePopupCloseClick = this._handlePopupCloseClick.bind(this);
     this._handleEscKeyDown = this._handleEscKeyDown.bind(this);
-
-    this._handleFavoriteClick = this._handleFavoriteClick.bind(this);
-    this._handleWatchedClick = this._handleWatchedClick.bind(this);
-    this._handleWatchList = this._handleWatchList.bind(this);
   }
 
-  // init(film, isFeaturesRerender) {
   init(film) {
     this._film = film;
     this._popupState.open = true;
 
-    this._popupComponent = new PopupView(this._film);
-
     document.addEventListener(`keydown`, this._handleEscKeyDown);
+
+    this._popupComponent = new PopupView(this._film);
 
     this._popupComponent.setPopupCloseClickHandler(this._handlePopupCloseClick);
 
-    this._popupComponent.setFavoriteClickHandler(this._handleFavoriteClick);
-    this._popupComponent.setWatchedClickHandler(this._handleWatchedClick);
-    this._popupComponent.setWatchlistClickHandler(this._handleWatchList);
-
     if (this._prevPopupComponent === null) {
-      render(this._popupContainerElement, this._popupComponent, RenderPosition.BEFOREEND);
-      this._renderCommentsBlock(this._film);
 
+      render(this._popupContainerElement, this._popupComponent, RenderPosition.BEFOREEND);
+
+      this._renderFeaturesBlock();
+      this._renderCommentsBlock();
       this._renderNewCommentBlock();
       this._prevPopupComponent = this._popupComponent;
-    // } else if (isFeaturesRerender) {
-    //   console.log(`asdf`);
+
     } else {
       remove(this._prevPopupComponent);
       render(this._popupContainerElement, this._popupComponent, RenderPosition.BEFOREEND);
 
-      this._renderCommentsBlock(this._film);
+      this._renderFeaturesBlock();
+      this._renderCommentsBlock();
       this._renderNewCommentBlock();
       this._prevPopupComponent = this._popupComponent;
     }
   }
 
-  _renderCommentsBlock(film) {
+  _renderFeaturesBlock() {
+    const featuresPresenter = new PopupFeaturesPresenter(
+        this._popupComponent.getFeaturesContainerElement(),
+        this._handleChangeData
+    );
+
+    featuresPresenter.init(this._film);
+  }
+
+  _renderCommentsBlock() {
     const commentsPresenter = new CommentsPresenter(this._popupComponent.getCommentsTitleElement());
 
-    commentsPresenter.init(film);
+    commentsPresenter.init(this._film);
   }
 
   _renderNewCommentBlock() {
-    if (!this._isNewCommentRendered) {
-      this._newCommentPresenter = new NewCommentPresenter(this._popupComponent.getCommentsWrapElement());
-    }
+    this._newCommentPresenter = new NewCommentPresenter(this._popupComponent.getCommentsWrapElement());
     this._newCommentPresenter.init();
-  }
-
-  _handleFeaturesClick(updatedData) {
-    const newData = Object.assign({}, this._film, updatedData);
-    this._handleChangeData(newData);
-    this.init(newData, true);
-  }
-
-  _handleFavoriteClick() {
-    this._handleFeaturesClick({isFavorite: !this._film.isFavorite});
-  }
-
-  _handleWatchedClick() {
-    this._handleFeaturesClick({isWatched: !this._film.isWatched});
-  }
-
-  _handleWatchList() {
-    this._handleFeaturesClick({isInWatchlist: !this._film.isInWatchlist});
   }
 
   _handlePopupCloseClick() {
@@ -102,7 +86,6 @@ export default class PopupPresenter {
     if (evt.key === `Escape` || evt.key === `Esc`) {
       evt.preventDefault();
       document.body.classList.remove(`hide-overflow`);
-      // this._popupComponent.reset(this._data);
       remove(this._popupComponent);
       this._popupState.open = false;
       document.removeEventListener(`keydown`, this._handleEscKeyDown);
